wip: move discard back to sigverify
