+++ perf/Cargo.toml
@@ -17,6 +17,7 @@ curve25519-dalek = { version = "3" }
 dlopen = "0.1.8"
 dlopen_derive = "0.1.4"
 fnv = "1.0.7"
+itertools = "0.10.3"
 lazy_static = "1.4.0"
 log = "0.4.14"
 rand = "0.7.0"
